Issue #7562 binary compatibility changes
