+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/OWALPage.java
@@ -1,121 +1,29 @@
-/*
- * Copyright 2010-2013 Orient Technologies LTD (info--at--orientechnologies.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package com.orientechnologies.orient.core.storage.impl.local.paginated.wal;
 
 import com.orientechnologies.common.serialization.types.OIntegerSerializer;
 import com.orientechnologies.common.serialization.types.OLongSerializer;
 import com.orientechnologies.orient.core.config.OGlobalConfiguration;
 
-import java.nio.ByteBuffer;
-
 /**
- * WAL page is organized using following format:
+ * Basic interface for classes which present pages of WAL.
+ * Main reason of creation of this interface is support of different binary formats
+ * of WAL in the same deployment.
  * <p>
- * <ol>
- * <li>CRC32 code of page content, it is used to check whether data are broken on disk. 4 bytes</li>
- * <li>Magic number, randomly generated number which is used to check whether page is broken on disk. 8 bytes</li>
- * <li>Amount of free space left on page, which can be used to add new records, 4 bytes</li>
- * <li>Position of LSN of last record which end is stored on this page, 8 bytes</li>
- * <li>End of the last record stored in page, 4 bytes</li>
- * <li>WAL records</li>
- * </ol>
- * <p>
- * Each WAL record is stored using following format:
- * <ol>
- * <li>Flag which indicates that record should be merged with record which is stored on next page. That is needed if
- * record can not be stored on one page and is split by two pages. 1 byte</li>
- * <li>Flag which indicates that this record is actually tail of long record parts of which are stored on other pages, 1 byte</li>
- * <li>Length of serialized content of WAL record. 4 bytes</li>
- * <li>Serialized content of the WAL record. Variable size.</li>
- * </ol>
- * <p>
- *
- * Every time new record is added. Value of free space left on page is updated.
- *
- * @author Andrey Lomakin
- * @since 5/8/13
+ * To detect which version of binary format is stored we use value stored under
+ * {@link OWALPage#MAGIC_NUMBER_OFFSET}
  */
-public class OWALPage {
-  static final        long MAGIC_NUMBER    = 0xFACB03FEL;
-  public static final int  PAGE_SIZE       = OGlobalConfiguration.DISK_CACHE_PAGE_SIZE.getValueAsInteger() * 1024;
-  static final        int  MIN_RECORD_SIZE = OIntegerSerializer.INT_SIZE + 3;
 
-  private static final int CRC_OFFSET          = 0;
-  static final         int MAGIC_NUMBER_OFFSET = CRC_OFFSET + OIntegerSerializer.INT_SIZE;
-  static final         int FREE_SPACE_OFFSET   = MAGIC_NUMBER_OFFSET + OLongSerializer.LONG_SIZE;
+public interface OWALPage {
+  int PAGE_SIZE       = OGlobalConfiguration.DISK_CACHE_PAGE_SIZE.getValueAsInteger() * 1024;
+  int MIN_RECORD_SIZE = OIntegerSerializer.INT_SIZE + 3;
 
-  /**
-   * Information about position of LSN of last record which end is stored on this page. If only begging of the record is stored on
-   * this page its position is not stored.
-   *
-   * @see OLogSegment.FlushTask#commitLog()
-   * @see OLogSegment#init(ByteBuffer)
-   */
-  static final int LAST_STORED_LSN = FREE_SPACE_OFFSET + OIntegerSerializer.INT_SIZE;
+  int CRC_OFFSET          = 0;
+  int MAGIC_NUMBER_OFFSET = CRC_OFFSET + OIntegerSerializer.INT_SIZE;
+  int FREE_SPACE_OFFSET   = MAGIC_NUMBER_OFFSET + OLongSerializer.LONG_SIZE;
 
-  /**
-   * End of the last record stored in page. This value is used in case of end of the WAL is broken and we want to truncate it to
-   * read only valid values
-   *
-   * @see OLogSegment.FlushTask#commitLog()
-   * @see OLogSegment#init(ByteBuffer)
-   */
-  static final int END_LAST_RECORD = LAST_STORED_LSN + OLongSerializer.LONG_SIZE;
+  byte[] getRecord(int position);
 
-  static final int RECORDS_OFFSET = END_LAST_RECORD + OIntegerSerializer.INT_SIZE;
+  boolean mergeWithNextPage(int position);
 
-  static final int MAX_ENTRY_SIZE = PAGE_SIZE - RECORDS_OFFSET;
-
-  private final ByteBuffer buffer;
-
-  OWALPage(ByteBuffer buffer, boolean isNew) {
-    this.buffer = buffer;
-
-    if (isNew) {
-      buffer.position(MAGIC_NUMBER_OFFSET);
-
-      buffer.putLong(MAGIC_NUMBER);
-      buffer.putInt(MAX_ENTRY_SIZE);
-
-      buffer.putLong(-1);// -1 means that we do not store any log record on this page
-      buffer.putInt(-1);// -1 means that we do not store any log record on this page
-    }
-  }
-
-  public byte[] getRecord(int position) {
-    buffer.position(position + 2);
-    final int recordSize = buffer.getInt();
-    final byte[] record = new byte[recordSize];
-    buffer.get(record);
-    return record;
-  }
-
-  boolean mergeWithNextPage(int position) {
-    return buffer.get(position) > 0;
-  }
-
-  int getFreeSpace() {
-    return buffer.getInt(FREE_SPACE_OFFSET);
-  }
-
-  static int calculateSerializedSize(int recordSize) {
-    return recordSize + OIntegerSerializer.INT_SIZE + 2;
-  }
-
-  static int calculateRecordSize(int serializedSize) {
-    return serializedSize - OIntegerSerializer.INT_SIZE - 2;
-  }
+  int getFreeSpace();
 }
