+++ core/src/test/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/OWALPageV2ChangesPortionTest.java
@@ -13,7 +13,7 @@ import java.util.Random;
  * @since 8/19/2015
  */
 @Test
-public class OWALPageChangesPortionTest {
+public class OWALPageV2ChangesPortionTest {
 
   @BeforeMethod
   public void before() {
